reimplement map panel (minimal)
