--- tools/twix-legacy/src/panels/map/layers/mod.rs
@@ -1,37 +1,0 @@
-mod ball_filter;
-mod ball_percepts;
-mod ball_position;
-mod ball_search_heatmap;
-mod behavior_simulator;
-mod field;
-mod image_segments;
-mod line_correspondences;
-mod lines;
-mod localization;
-mod obstacle_filter;
-mod obstacles;
-mod path;
-mod path_obstacles;
-mod pose_detection;
-mod referee_position;
-mod robot_pose;
-mod voronoi_cell;
-
-pub use ball_filter::BallFilter;
-pub use ball_percepts::BallPercepts;
-pub use ball_position::BallPosition;
-pub use ball_search_heatmap::BallSearchHeatmap;
-pub use behavior_simulator::BehaviorSimulator;
-pub use field::Field;
-pub use image_segments::ImageSegments;
-pub use line_correspondences::LineCorrespondences;
-pub use lines::Lines;
-pub use localization::Localization;
-pub use obstacle_filter::ObstacleFilter;
-pub use obstacles::Obstacles;
-pub use path::Path;
-pub use path_obstacles::PathObstacles;
-pub use pose_detection::PoseDetection;
-pub use referee_position::RefereePosition;
-pub use robot_pose::RobotPose;
-pub use voronoi_cell::VoronoiCell;
