+++ tools/twix/Cargo.toml
@@ -6,8 +6,10 @@ license.workspace = true
 homepage.workspace = true
 
 [dependencies]
+behavior_node = { workspace = true }
 clap = { workspace = true }
 color-eyre = { workspace = true }
+convert_case = { workspace = true }
 coordinate_systems = { workspace = true }
 dirs = { workspace = true }
 eframe = { workspace = true }
@@ -19,6 +21,7 @@ image = { workspace = true }
 itertools = { workspace = true }
 linear_algebra = { workspace = true }
 log = { workspace = true }
+nalgebra = { workspace = true }
 projection = { workspace = true }
 repository = { workspace = true }
 ros-z = { workspace = true }
@@ -32,3 +35,4 @@ toml = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
 types = { workspace = true }
 uuid = { workspace = true }
+voronoi = { workspace = true }
