+++ tools/twix/src/panels/map/layers/field.rs
@@ -3,14 +3,14 @@ use std::sync::Arc;
 use color_eyre::Result;
 use types::field_dimensions::FieldDimensions;
 
-use crate::{panels::map::layer::Layer, robot::Robot, twix_painter::TwixPainter};
+use crate::{backend::RobotBackend, panels::map::layer::Layer, twix_painter::TwixPainter};
 
 pub struct Field {}
 
 impl Layer<coordinate_systems::Field> for Field {
     const NAME: &'static str = "Field";
 
-    fn new(_robot: Arc<Robot>) -> Self {
+    fn new(_backend: Arc<RobotBackend>) -> Self {
         Self {}
     }
