+++ tools/twix/src/panels/map/layers/image_segments.rs
