+++ tools/twix/src/panels/map/layers/obstacles.rs
