+++ tools/twix/src/panels/mod.rs
@@ -1,5 +1,7 @@
 mod image;
+mod map;
 mod text;
 
 pub use image::ImagePanel;
+pub use map::MapPanel;
 pub use text::TextPanel;
