+++ tools/twix/src/zoom_and_pan.rs
